perf(backend): Converted theme switch to preact component as well.
My my was this a pain. 😭

Changelog:

Moved everything to react component

Changed the internal css to external sass

Added visibile client directive



= Brekaing Changes =

sass takes a while to load and make moon invisible as the script has to aadd the default dark/light flag

= Other notes =
If  you plan to add svvg into preact, you can just copy and paste without worrying. It is made to be compatible.